Moved extension classes to seperate files as Jimmacle requires.
